first implementation for the schwartz plugin
