+++ README.md
@@ -1,18 +1,20 @@
-#Quartz Monitor Plugin for Grails
+#Schwartz Monitor Plugin for Grails
 
-Allows you to view and administer all your Quartz jobs in one place.
+This plugin is a fork of the [quartz-monitor](https://grails.org/plugin/quartz-monitor) plugin and supports the Grails [Schwartz](https://plugins.grails.org/plugin/schwartz) plugin instead of grails-quartz.
+
+It allows you to view and administer all your Quartz job services in the web-ui.
 
 ##Prerequisites
 
-This plugin requires the [Quartz](http://grails.org/plugin/quartz) and [Asset Pipeline](http://grails.org/plugin/asset-pipeline) plugins to run.
+This plugin requires the Grails [Schwartz](https://plugins.grails.org/plugin/schwartz) and [Asset Pipeline](http://grails.org/plugin/asset-pipeline) plugins to run.
 
 ##Usage
 
-Once you have the Quartz plugin installed and have created some jobs, then you will probably start wondering if they are all running as desired.
+Once you have the Schwartz plugin installed and have created some job services, then you will probably start wondering if they are all running as desired.
 
-This is where you need the 'Quartz Monitor' plugin.
+This is where you need the 'Schwartz Monitor' plugin.
 
-Simply install the plugin and go to the URL: http://localhost:8080/<yourapp>/quartz and you will find all a list of all the jobs you have scheduled to run.
+Simply install the plugin and go to the URL: http://localhost:8080/<yourapp>/quartz and you will find all a list of all the Quartz job services you have created.
 
 ##Enhanced Experience
 
@@ -38,3 +40,6 @@ Will add javascript to the page in order to show a countdown to when the job wil
 
 Will add javascript to the page in order to show a clock with the current time, unless this is set to 'false'.
 
+## Internals
+
+This plugin replaces the Schwartz "quartzJobFactory" bean with an own one, which wraps the QuartzMonitorJobFactory. This allows the plugin to display the lastRun date, which is not available in the Schwartz plugin currently.
