+++ classes/artifacts/test-harness/exploded/test-harness-0.1.war/META-INF/MANIFEST.MF
@@ -1,0 +1,2 @@
+Manifest-Version: 1.0
+
