+++ gradle.properties
@@ -1,2 +1,2 @@
-grailsVersion=3.0.4
-gradleWrapperVersion=2.9
+grailsVersion=3.0.17
+gradleWrapperVersion=2.3
