--- grails-app/assets/javascripts/quartz-monitor.js
@@ -1,1 +1,0 @@
-//=require quartz-monitor-main
