+++ grails-app/assets/javascripts/schwartz-monitor-main.js
