+++ grails-app/assets/javascripts/schwartz-monitor.js
@@ -1,0 +1,1 @@
+//=require schwartz-monitor-main
