--- grails-app/assets/stylesheets/quartz-monitor.css
@@ -1,3 +1,0 @@
-/*
-*= require quartz-monitor-main
-*/
