+++ grails-app/assets/stylesheets/schwartz-monitor-main.css
