+++ grails-app/assets/stylesheets/schwartz-monitor.css
@@ -1,0 +1,3 @@
+/*
+*= require schwartz-monitor-main
+*/
