--- src/test/projects/test-harness-no-hibernate/grails-app/jobs/test/harness/NoTriggersJob.groovy
@@ -1,12 +1,0 @@
-package test.harness
-
-class NoTriggersJob {
-    def group = "MyGroup"
-
-    static triggers = {
-    }
-
-    def execute() {
-        println "running NoTriggersJob"
-    }
-}
