--- src/test/projects/test-harness-no-hibernate/grails-app/jobs/test/harness/TestJob.groovy
@@ -1,12 +1,0 @@
-package test.harness
-
-
-class TestJob {
-    static triggers = {
-        simple repeatInterval: 50000
-    }
-
-    def execute() {
-        println "running TestJob"
-    }
-}
