+++ src/test/projects/test-harness/gradle.properties
@@ -1,2 +1,2 @@
-grailsVersion=3.0.4
+grailsVersion=3.0.17
 gradleWrapperVersion=2.9
