--- src/test/projects/test-harness/grails-app/jobs/test/harness/CronJob.groovy
@@ -1,12 +1,0 @@
-package test.harness
-
-
-class CronJob {
-    static triggers = {
-        cron startDelay:2000, cronExpression: '0 * * * * ?'
-    }
-
-    def execute() {
-        println "running CronJob"
-    }
-}
