--- src/test/projects/test-harness/grails-app/jobs/test/harness/TestJob.groovy
@@ -1,14 +1,0 @@
-package test.harness
-
-
-class TestJob {
-    def description = ''
-
-    static triggers = {
-        simple repeatInterval: 50000
-    }
-
-    def execute() {
-        println "running TestJob"
-    }
-}
